Create CSC667
Introduction, Problem, and Objectives

One of the most common cancers among women is Breast Cancer, it occurs when breast cells mutate and grow out of control, creating a mass of tissue called a Tumor. Research shows different kinds of breast tumors, but we will focus on Benign and Malignant tumors.

Benign tumors are non-cancerous and not life-threatening. Malignant breast tumors, however, if not detected and treated early, will continue to grow, invading and destroying adjacent normal tissue, cancer cells will break away from the tumor and spread through the lymph system and bloodstream. At this stage, breast cancer is deadly, and chances of cure are far lower than in the early, localized stage. Research shows that breast cancer is the second leading cause of cancer death among women. It’s most likely to affect women over the age of 50. Breast cancer is sometimes found after symptoms appear, but many women with breast cancer have no symptoms.

To prevent breast cancer death, early detection is necessary to save lives. According to the American Cancer Society, when breast cancer is detected early, and is in the localized stage, the 5-year relative survival is 99%.

In our cause of study, Transfer Learning is applied in the detection of breast cancer using various pre-trained models. Transfer Learning is the process of using pre-trained models that were used to solve one problem and using that knowledge to solve other particular problems. Transfer Learning is an approach where we use one model trained on a machine learning task and reuse it as a starting point for a different job. Transfer Learning is easy, faster, and more accurate compared to other traditional models, it gives us the ability to share learned features across different learning tasks.
Our study objectives are:
* Detecting Malignant tumors at an early stage based on Ultrasound images can increase the chances of better treatment.
* Predicting the tumorous cells using transfer learning models.